requirements.txt
requests>=2.31.0
pandas>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pycountry>=23.12.0
python-dotenv>=1.0.0

# Frontend
streamlit>=1.32.0
pydeck

# Testing
pytest>=8.0.0
requests-mock>=1.11.0
pytest-cov
